fix teleport desync
