+++ pumpkin-protocol/src/client/play/mod.rs
@@ -68,6 +68,7 @@ mod spawn_entity;
 mod stop_sound;
 mod store_cookie;
 mod subtitle;
+mod synchronize_vehicle_position;
 mod system_chat_message;
 mod take_item;
 mod teleport_entity;
@@ -151,6 +152,7 @@ pub use spawn_entity::*;
 pub use stop_sound::*;
 pub use store_cookie::*;
 pub use subtitle::*;
+pub use synchronize_vehicle_position::*;
 pub use system_chat_message::*;
 pub use take_item::*;
 pub use teleport_entity::*;
